Removed all Date and Calendar instances replaces with LocalDate and LocalTime. Removed all the Date Instances with LocalDate / LocalTime As with Dates it need to used with Calendar Api in most of the cases , LocalDate and others have direct methods for adding hours/ days ... etc. Thus removed all the Calendar Instances.
Core Library de-sugaring is enabled which let use this Java8 App's. Added de-sugaring library.

Converters as the Date was converted to millisSinceEpoch thus Its changed such a way that the LocalDate is converted from millis and LocalDate to converted to millis thus no need to add a migrations.

In DateExtension.kt replaced SimpleDateFormat with DateTimeFormatter this API is thread safe and recommended to use.

Some changes are added those aren't't related to Date formatting like the use of Recurrence enum class rather than using the string values.

There is some confusion in some places weather to use LocalTime or LocalDateTime like in Medication::medicationTime used LocalDateTime for the time being

StateSavers holds the Savers for the non-primitive data type for the AddMedicationRoute text fields.

Recurrence getRecurrenceList is removed as all the Recurrence value can be got by .entries

The other changes is same Date -> LocalDate or LocalDateTime and less use of Calendar Api.